Created Data Manager, dynamically updated time
The Data Manager was created to centralize CSV reading and data storage. This was used as a bridge between the raw CSV file and other two managers. Moreover, the algorithm for updating the time was created, and some code was refactored. A few bug fixes were done thanks to the new Data Manager.